refactor: SM-6 컴포넌트 모듈화 및 코드 리팩토링
