+++ src/models/types.ts
@@ -1,0 +1,3 @@
+type SizeType = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+
+export type { SizeType };
