--- src/types/index.ts
@@ -1,5 +1,0 @@
-export type SizeType = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
-
-export type ButtonSizeType = Extract<SizeType, 'sm' | 'md' | 'lg'>;
-
-export type ButtonColorType = 'dark' | 'primary' | 'light';
